Moving the fixtures to conftest file
Moving the fixtures from test_for_review_course to the conftest.py
